move services store to service.ts
